fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@11247 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/activitygen/AGActivityGenHandler.cpp
@@ -197,8 +197,8 @@ AGActivityGenHandler::parseCityGates(const SUMOSAXAttributes &attrs) {
 		string edge = attrs.getString(SUMO_ATTR_EDGE);
 		SUMOReal positionOnEdge = attrs.getFloat(SUMO_ATTR_POSITION);
 		AGPosition posi(myCity.getStreet(edge), positionOnEdge);
-		myCity.statData.incoming[myCity.cityGates.size()] = attrs.getFloat(AGEN_ATTR_INCOMING);
-		myCity.statData.outgoing[myCity.cityGates.size()] = attrs.getFloat(AGEN_ATTR_OUTGOING);
+		myCity.statData.incoming[(int)myCity.cityGates.size()] = attrs.getFloat(AGEN_ATTR_INCOMING);
+		myCity.statData.outgoing[(int)myCity.cityGates.size()] = attrs.getFloat(AGEN_ATTR_OUTGOING);
 		myCity.cityGates.push_back(posi);
 
 	} catch (const exception& e) {
